Rework the In-call volume management.
The in-call volume management is managed as a specific mixer with its specific alsa properties (alsa_incall_vol_properties_t). This is in line with the spirit of alsa_sound libaudio.
This property array is used to identify the widgets needed to change the modem downlink audio stream volume for the supported audio devices.
Today, the devices supported in voice call are EARPIECE, SPEAKER, HEADSET, and BLUETOOTH SCO.

When the in-call volume method setVoiceVolume is called all volume gain of all supported devices are changed.

Thanks to this change, in-call volume can be supported in voice call Bluetooth also.

All code changes are under AUDIO_MODEM_TI flag.

Change-Id: I2d95e4af0f1e202660abda63781f4618a72e4731